Update preonic layout for preonic keyboard
I don't really use this keyboard much anymore because the Atreus is so
good, but am giving it another shot because with the 2 unit spacebars
it's much more comfortable than before and my keystrokes are much more
confident.

Significant changes to the layout because the previous one just wasn't
that good. I'm coming to realize that the 40% layout really is the
layout for me, as it forces me to make difficult decisions that end up
greatly increasing the comfort level while typing.

The real problem with the bigger spacebars is they are so, so, so
sluggish and I'm finding even typing out this commit message that they
slow down my typing just a little. Maybe they're better without the
stabilizers? Maybe the stabilizers need to be lubed? No idea, but it's
not a good feeling.