Allow local language servers
This allows for local executables that map to language servers to be
enabled in addition to Mason-managed language servers. This lets me use
local versions of ruff if it's installed, or to not use it if it's not
(and hopefully use something like black if not). Except for NPM
projects, I am in the habit of adding local environments to my PATH, so
Python packages are usually available as executables.

There is still some weirdness here: there are language server configs
that are functions and make it hard to figure out how to point things at
node_modules or a local executable.